tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.attributes.shadowColor.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.attributes.shadowColor.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.attributes.shadowColor.invalid</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.shadowColor = '#00ff00';
  ctx.shadowColor = 'bogus';
  _assertSame(ctx.shadowColor, '#00ff00', "ctx.shadowColor", "'#00ff00'");

  ctx.shadowColor = '#00ff00';
  ctx.shadowColor = 'red bogus';
  _assertSame(ctx.shadowColor, '#00ff00', "ctx.shadowColor", "'#00ff00'");

  ctx.shadowColor = '#00ff00';
  ctx.shadowColor = ctx;
  _assertSame(ctx.shadowColor, '#00ff00', "ctx.shadowColor", "'#00ff00'");

  ctx.shadowColor = '#00ff00';
  ctx.shadowColor = undefined;
  _assertSame(ctx.shadowColor, '#00ff00', "ctx.shadowColor", "'#00ff00'");
  t.done();

});
</script>
